Refactored test helpers files into one model.
Using a model is preferable because inheritance can be used for appropriate overriding in each module.
