+++ nh_eobs/__init__.py
@@ -17,3 +17,4 @@ from . import wardboard
 from . import wizard
 from . import workload
 from .models import nh_clinical_patient_monitoring_exception
+from .tests.common import nh_clinical_test_utils
