+++ nh_eobs/tests/nh_eobs_api/test_placement.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from openerp.addons.nh_eobs.tests.common import test_data_creator
 from openerp.tests.common import TransactionCase
 
 
@@ -7,25 +6,30 @@ class TestPlacement(TransactionCase):
 
     def setUp(self):
         super(TestPlacement, self).setUp()
-        test_data_creator.admit_patient(self)
+        self.test_utils_model = self.env['nh.clinical.test_utils']
+        self.test_utils_model.admit_patient()
+
+        self.spell_activity_id = self.test_utils_model.spell_activity_id
         self.domain = [
             ('data_model', '=', 'nh.clinical.patient.observation.ews'),
             ('spell_activity_id', '=', self.spell_activity_id),
             ('state', 'not in', ['completed', 'cancelled'])
         ]
+        self.activity_model = self.env['nh.activity']
+        self.api_model = self.env['nh.eobs.api']
 
     def test_new_obs_created(self):
         # Only cancel when transfer successful.
         open_obs_activities = self.activity_model.search(self.domain)
         self.assertEqual(len(open_obs_activities), 0)
 
-        test_data_creator.place_patient(self)
+        self.test_utils_model.place_patient()
 
         open_obs_activities = self.activity_model.search(self.domain)
         self.assertEqual(len(open_obs_activities), 1)
 
     def test_new_obs_due_in_15_minutes(self):
-        test_data_creator.place_patient(self)
+        self.test_utils_model.place_patient()
 
         open_obs_activities = self.activity_model.search(self.domain)
         self.assertEqual(len(open_obs_activities), 1)
@@ -33,7 +37,7 @@ class TestPlacement(TransactionCase):
         self.assertEqual(open_obs_activity.data_ref.frequency, 15)
 
     def test_new_obs_due_in_15_minutes_after_transfer(self):
-        test_data_creator.place_patient(self)
+        self.test_utils_model.place_patient()
         # open_obs_activities = self.activity_model.search(self.domain)
         # first_obs_after_placement = open_obs_activities[0]
         # TODO Uncomment 2 lines above when EOBS-690 has been completed.
@@ -46,7 +50,7 @@ class TestPlacement(TransactionCase):
         # TODO Uncomment line above when EOBS-690 has been completed.
 
         # Place patient again on new ward.
-        test_data_creator.place_patient(self)
+        self.test_utils_model.place_patient()
 
         open_obs_activities = self.activity_model.search(self.domain)
         self.assertEqual(len(open_obs_activities), 1)
