+++ nh_eobs/tests/nh_eobs_api/test_transfer.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from openerp.addons.nh_eobs.tests.common import test_data_creator
 from openerp.tests.common import TransactionCase
 
 
@@ -7,7 +6,8 @@ class TestTransfer(TransactionCase):
 
     def setUp(self):
         super(TestTransfer, self).setUp()
-        test_data_creator.admit_and_place_patient(self)
+        self.test_utils_model = self.env['nh.clinical.test_utils']
+        self.test_utils_model.admit_and_place_patient()
         self.domain = [
             ('data_model', '=', 'nh.clinical.patient.observation.ews'),
             ('spell_activity_id', '=', self.spell_activity_id),
