+++ nh_eobs/tests/test_get_last_full_obs_activity.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from openerp.addons.nh_eobs.tests.common import test_data_creator
 from openerp.addons.nh_ews.tests.common import clinical_risk_sample_data
 from openerp.tests.common import TransactionCase
 
@@ -8,9 +7,10 @@ class TestGetLastFullObs(TransactionCase):
 
     def setUp(self):
         super(TestGetLastFullObs, self).setUp()
-        test_data_creator.admit_and_place_patient(self)
+        self.test_utils_model = self.env['nh.clinical.test_utils']
+        self.test_utils_model.admit_and_place_patient()
 
-        self.observation_test_utils = self.env['observation_test_utils']
+        self.test_utils_model = self.env['nh.clinical.test_utils']
 
         self.domain = [
             ('data_model', '=', 'nh.clinical.patient.observation.ews'),
@@ -20,7 +20,7 @@ class TestGetLastFullObs(TransactionCase):
 
     def test_returns_full_obs(self):
         self.initial_medium_risk_obs_activity = \
-            self.observation_test_utils.create_ews_obs_activity(
+            self.test_utils_model.create_ews_obs_activity(
                 self.patient.id, self.spell_activity_id,
                 clinical_risk_sample_data.MEDIUM_RISK_DATA
             )
@@ -38,13 +38,13 @@ class TestGetLastFullObs(TransactionCase):
     def test_returns_obs_from_before_transfer_after_placement(
             self):
         self.initial_medium_risk_obs_activity = \
-            self.observation_test_utils.create_ews_obs_activity(
+            self.test_utils_model.create_ews_obs_activity(
                 self.patient.id, self.spell_activity_id,
                 clinical_risk_sample_data.MEDIUM_RISK_DATA
             )
 
         self.api_model.transfer(self.hospital_number, {'location': 'WB'})
-        test_data_creator.place_patient(self)
+        self.test_utils_model.place_patient()
 
         obs_activity = \
             self.ews_model.get_last_full_obs_activity(self.spell_activity_id)
@@ -54,16 +54,16 @@ class TestGetLastFullObs(TransactionCase):
     def test_returns_obs_from_before_transfer_after_placement_and_refusal(
             self):
         self.initial_medium_risk_obs_activity = \
-            self.observation_test_utils.create_ews_obs_activity(
+            self.test_utils_model.create_ews_obs_activity(
                 self.patient.id, self.spell_activity_id,
                 clinical_risk_sample_data.MEDIUM_RISK_DATA
             )
 
         self.api_model.transfer(self.hospital_number, {'location': 'WB'})
-        test_data_creator.place_patient(self)
+        self.test_utils_model.place_patient()
 
         self.partial_obs_activity = \
-            self.observation_test_utils.refuse_open_obs(
+            self.test_utils_model.refuse_open_obs(
                 self.patient.id, self.spell_activity_id
             )
