+++ nh_eobs_mental_health/__init__.py
@@ -10,5 +10,6 @@ from .models import nh_clinical_wardboard
 from .models import nh_eobs_api
 from .models import nh_eobs_mobile_main
 from .models import nh_eobs_ward_dashboard
+from .tests.common import nh_clinical_test_utils
 from .wizard import nh_clinical_patient_monitoring_exception_select_reason
