+++ nh_ews/__init__.py
@@ -2,3 +2,4 @@
 from . import ews
 from . import notifications
 from . import parameters
+from .tests.common import nh_clinical_test_utils
